Support multiple containers and volumes
It is now possible to inject multiple containers and volumes into a
single pod. For this we renamed the annotation for the container and
volume name and it is now possible to provide a comma-separated list of
container and volume names via the corresponding annotations.

Environment variables are still supported, but now we must also provide
the container for which the environment variable should be used in the
configuration file.