:warning: Change Logging Bahaviour (#19)
Instead of using a custom "log" package, we are now using the log
package of the controller runtime. This means that we had to remove our
custom "--log.level" and "--log.format" flags and replaced them with the
built-in zap flags (e.g. "--zap-log-level" and "--zap-encoder").

NOTE: This is a breaking change, which also required some adjustments in
the Helm chart, so that the new Helm chart version isn't working with
the old Docker images. This means if the Helm chart is updated to
version 0.8.0 also the v0.8.0 Docker image must be used.