Add `injectors` Configuration
Until now it was only possible to inject sidecars into a Pod, when the
Pod had the `sidecar-injector.ricoberger.de` annotation. With the new
`injectors` configuration it is now possible to inject sidecars by
defining them in the configuration. When defined the sidecars will be
injected into all Pods which are matching the `selectors` configuration,
so that the Pod must not have the `sidecar-injector.ricoberger.de`
annotation anymore.